+++ packages/carbon-blocks-react/src/components/BlockOptions/OptionButton.tsx
@@ -1,0 +1,21 @@
+import { preventAndStop } from "@emrgen/carbon-core";
+import { ReactNode } from "react";
+
+interface OptionMenuProps {
+  children: ReactNode;
+  onClick: (e) => void;
+}
+
+export const OptionButton = (props: OptionMenuProps) => {
+  const { children, onClick } = props;
+
+  return (
+    <div
+      className={"block-option-button"}
+      onClick={onClick}
+      onMouseDown={preventAndStop}
+    >
+      {children}
+    </div>
+  );
+};