+++ packages/carbon-blocks-react/src/components/BlockOptions/mod.ts
@@ -1,0 +1,2 @@
+export * from './OptionMenu'
+export * from './BlockOptions'