+++ apps/react/CHANGELOG.md
@@ -8,7 +8,7 @@
   - @emrgen/carbon-core@0.0.5
   - @emrgen/carbon-clipboard@0.0.5
   - @emrgen/carbon-react@0.0.4
-  - @emrgen/carbon-react-blocks@0.0.4
+  - @emrgen/carbon-blocks-react@0.0.4
   - @emrgen/carbon-utils@0.0.4
   - @emrgen/fastype-core@0.0.4
 
@@ -17,7 +17,7 @@
 ### Patch Changes
 
 - Updated dependencies [235e167]
-  - @emrgen/carbon-react-blocks@0.0.3
+  - @emrgen/carbon-blocks-react@0.0.3
   - @emrgen/carbon-sheet@0.0.3
   - @emrgen/carbon-utils@0.0.3
   - @emrgen/fastype-core@0.0.3
@@ -30,7 +30,7 @@
 ### Patch Changes
 
 - Updated dependencies [235e167]
-  - @emrgen/carbon-react-blocks@0.0.3-next.0
+  - @emrgen/carbon-blocks-react@0.0.3-next.0
   - @emrgen/carbon-sheet@0.0.3-next.0
   - @emrgen/carbon-utils@0.0.3-next.0
   - @emrgen/fastype-core@0.0.3-next.0